Moved RecipientKind to utils from mobile
