+++ packages/mobile/src/send/Send.tsx
@@ -1,6 +1,7 @@
 import colors from '@celo/react-components/styles/colors'
 import { fontStyles } from '@celo/react-components/styles/fonts'
 import { componentStyles } from '@celo/react-components/styles/styles'
+import { RecipientKind } from '@celo/utils/src/recipient'
 import { throttle } from 'lodash'
 import * as React from 'react'
 import { withNamespaces, WithNamespaces } from 'react-i18next'
@@ -27,7 +28,6 @@ import {
   filterRecipients,
   NumberToRecipient,
   Recipient,
-  RecipientKind,
   RecipientWithQrCode,
 } from 'src/utils/recipient'
