+++ packages/mobile/src/send/SendAmount.tsx
@@ -5,6 +5,7 @@ import colors from '@celo/react-components/styles/colors'
 import { fontStyles } from '@celo/react-components/styles/fonts'
 import { componentStyles } from '@celo/react-components/styles/styles'
 import { parseInputAmount } from '@celo/utils/src/parsing'
+import { RecipientKind } from '@celo/utils/src/recipient'
 import BigNumber from 'bignumber.js'
 import { debounce } from 'lodash'
 import * as React from 'react'
@@ -46,7 +47,7 @@ import { ConfirmationInput } from 'src/send/SendConfirmation'
 import DisconnectBanner from 'src/shared/DisconnectBanner'
 import { fetchDollarBalance } from 'src/stableToken/actions'
 import Logger from 'src/utils/Logger'
-import { Recipient, RecipientKind } from 'src/utils/recipient'
+import { Recipient } from 'src/utils/recipient'
 
 const TAG: string = 'send/SendAmount'
