+++ packages/mobile/src/send/saga.test.ts
@@ -1,3 +1,4 @@
+import { RecipientKind } from '@celo/utils/src/recipient'
 import { expectSaga } from 'redux-saga-test-plan'
 import { select } from 'redux-saga/effects'
 import { showError } from 'src/alert/actions'
@@ -9,7 +10,6 @@ import { BarcodeTypes } from 'src/qrcode/utils'
 import { Actions, QrCode } from 'src/send/actions'
 import { recipientCacheSelector } from 'src/send/reducers'
 import { watchQrCodeDetections } from 'src/send/saga'
-import { RecipientKind } from 'src/utils/recipient'
 import { mockAccount, mockE164Number, mockName, mockQrCodeData } from 'test/values'
 
 jest.mock('src/utils/time', () => ({
