+++ packages/mobile/test/values.ts
@@ -1,4 +1,5 @@
 /* Shared mock values to facilitate testing */
+import { RecipientKind } from '@celo/utils/src/recipient'
 import BigNumber from 'bignumber.js'
 import { MinimalContact } from 'react-native-contacts'
 import { NotificationTypes, PaymentRequest, PaymentRequestStatuses } from 'src/account/types'
@@ -6,7 +7,7 @@ import { EscrowedPayment } from 'src/escrow/actions'
 import { SHORT_CURRENCIES } from 'src/geth/consts'
 import { AddressToE164NumberType, E164NumberToAddressType } from 'src/identity/reducer'
 import { AttestationCode } from 'src/identity/verification'
-import { RecipientKind, RecipientWithContact, RecipientWithMobileNumber } from 'src/utils/recipient'
+import { RecipientWithContact, RecipientWithMobileNumber } from 'src/utils/recipient'
 
 export const mockName = 'John Doe'
 export const mockAccount = '0x0000000000000000000000000000000000007E57'
