+++ packages/utils/src/recipient.ts
@@ -1,0 +1,6 @@
+export enum RecipientKind {
+  MobileNumber = 'mobileNumber',
+  Contact = 'contact',
+  QrCode = 'QrCode',
+  Address = 'Address',
+}
